requirements.txt
starlette
uvicorn
requests
python-telegram-bot
plexapi